requirements.txt
cookiecutter==1.7.3
black==21.9b0
pyyaml==5.4.1
